feat: extract MobileFieldInput from MobileField component
so verified field can also use it
