+++ frontend/src/templates/Field/Mobile/types.ts
@@ -1,0 +1,3 @@
+import { FormFieldWithId, MobileFieldBase } from '~shared/types/field'
+
+export type MobileFieldSchema = FormFieldWithId<MobileFieldBase>
